Create Program2_2.c
This C program prints a line of asterisks (*) equal to a number entered by the user. The function Display(int iNo) uses a while loop that continues as long as iNo is greater than 0. In each iteration, it prints an asterisk and decrements iNo by 1. The main() function prompts the user to enter a number, which is then passed to Display(), resulting in a horizontal line of asterisks corresponding to the input value. This program demonstrates the use of a while loop for repetitive printing. The time complexity is O(n), where n is the value of the input number.