Create Program2_3.c
This C program prints either "Hello" or "Demo" based on the value entered by the user. The function Display(int iNo) checks if the input number is less than 10. If it is, the program prints "Hello"; otherwise, it prints "Demo". In the main() function, the user is prompted to enter a number, which is then passed to Display() to determine the output. This program demonstrates the use of conditional statements (if-else) to make decisions based on input. The time complexity is O(1), as it involves only a single comparison and print operation.