Create Program2_4.c
This C program prints a given number multiple times based on a user-specified frequency. The function Display(int iNo, int iFrequency) uses a for loop that iterates from 1 to iFrequency, printing the number iNo in each iteration. In the main() function, the user is prompted to enter a number and the number of times (frequency) it should be printed. These values are passed to Display(), which prints the number consecutively the specified number of times. This program demonstrates the use of loops and repeated output. The time complexity is O(n), where n is the value of iFrequency.