Create Program2_5.c
This C program checks whether a number entered by the user is even or odd. The function ChkEven(int iNo) takes an integer as input and returns true if the number is divisible by 2 (even), and false otherwise (odd). In the main() function, the user is prompted to enter a number, which is passed to ChkEven(). Based on the returned boolean value, the program prints either "Number is Even" or "Number is odd". This program demonstrates the use of conditional checks, functions, and boolean values. The time complexity is O(1) since it performs a single modulo operation and comparison.