Create Program4_5.c
This C program calculates the difference between the sum of factors and the sum of non-factors of a given number. The program defines a function FactDiff that takes an integer as input and first converts negative numbers to their absolute value. Inside the function, two variables are initialized to store the sum of factors (iSumFact) and the sum of non-factors (iSumNonFact). A loop iterates through all numbers from 1 up to, but not including, the input number. For each number, the program checks if it is a factor of the input (i.e., divides the number evenly). If it is a factor, it is added to iSumFact; otherwise, it is added to iSumNonFact. After completing the loop, the function returns the difference between the sum of factors and the sum of non-factors. In the main function, the program prompts the user to enter a number, calls FactDiff to calculate the result, and then prints the difference. The program has a time complexity of O(n), as it iterates through all numbers less than the input. For example, if the input is 6, the factors are 1, 2, 3 (sum = 6), and the non-factors are 4, 5 (sum = 9), so the output will be -3.