Create Program12_4.c
The sum_of_even_factors function calculates the sum of all even factors of a given number. It uses a for loop to iterate from 1 up to one less than the input number, checking each integer to see if it is both a factor of the number (number % i == 0) and even (i % 2 == 0). If both conditions are true, the integer is added to the cumulative sum sum. Since the loop examines every number from 1 to the input number minus one, the number of iterations grows linearly with the input value. Therefore, the execution time increases proportionally with the input, and the time complexity of this function is O(n), where n is the input number.